chore(loadgen): refactor firebase code into auth and client
